requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==20.3
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
